refactor: move x
