+++ tests/go/internal/x/xtest_blackbox_test.go
@@ -3,7 +3,7 @@ package x_test
 import (
 	"testing"
 
-	"github.com/fredrikaverpil/neotest-golang/x"
+	"github.com/fredrikaverpil/neotest-golang/internal/x"
 )
 
 func TestBlackBox(t *testing.T) {
